Also create submodule src/assets/ignored_suffixes.rs
